amount is decial

+++ flexibee-core/src/main/java/com/adleritech/flexibee/core/api/domain/IssuedInvoiceItem.java
@@ -21,7 +21,7 @@ public class IssuedInvoiceItem {
     String name;
 
     @Element(name = "mnozBaleni", required = false)
-    Integer amount;
+    BigDecimal amount;
 
     @Element(name = "szbDph", required = false)
     BigDecimal vatRate;
